Feat: CanvasChart 구현 위해 해당 로직울 component로 이동 - Chart 데이터 요청 및 canvas 생성 로직 이동
